reorg utils folder
