+++ ui/panels-plugin/src/plugins/line-chart/utils/focused-series.test.ts
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { getNearbySeries } from './tooltip-model';
+import { getNearbySeries } from './focused-series';
 
 describe('getNearbySeries', () => {
   const seriesMock = [
