+++ ui/panels-plugin/src/plugins/line-chart/utils/palette-gen.ts
